requirements.txt
pandas
ydata_profiling
great_expectations
psycopg2
python-dotenv